pyproject.toml
[project]
name = "enhanced-cfn-mcp-server"
version = "2.0.0"
description = "An enhanced AWS CloudFormation Model Context Protocol (MCP) server with intelligent prompting, autonomous fixing, and comprehensive AWS resource management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.0",
    "pydantic>=2.10.6",
    "mcp[cli]>=1.6.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "pyyaml>=6.0.0",
]
license = {text = "Apache-2.0"}
license-files = ["LICENSE", "NOTICE"]
authors = [
    {name = "Amazon Web Services"},
    {name = "Enhanced CFN MCP Contributors"},
]
keywords = [
    "aws",
    "cloudformation",
    "mcp",
    "model-context-protocol",
    "infrastructure-as-code",
    "amazon-q",
    "ai-assistant"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
homepage = "https://github.com/shantgup/enhanced-cfn-mcp-server"
documentation = "https://github.com/shantgup/enhanced-cfn-mcp-server#readme"
repository = "https://github.com/shantgup/enhanced-cfn-mcp-server"
"Bug Tracker" = "https://github.com/shantgup/enhanced-cfn-mcp-server/issues"
changelog = "https://github.com/shantgup/enhanced-cfn-mcp-server/blob/main/CHANGELOG.md"

[project.scripts]
"enhanced-cfn-mcp-server" = "awslabs.cfn_mcp_server.server:main"

[project.optional-dependencies]
dev = [
    "commitizen>=4.2.2",
    "pre-commit>=4.1.0",
    "ruff>=0.9.7",
    "pyright>=1.1.398",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["awslabs"]

[tool.ruff]
line-length = 99
extend-include = ["*.ipynb"]
exclude = [
    ".venv",
    "**/__pycache__",
    "**/node_modules",
    "**/dist",
    "**/build",
    "**/env",
    "**/.ruff_cache",
    "**/.venv",
    "**/.ipynb_checkpoints"
]
force-exclude = true

[tool.ruff.lint]
exclude = ["__init__.py"]
select = ["C", "D", "E", "F", "I", "W"]
ignore = ["C901", "E501", "E741", "F402", "F823", "D100", "D106"]

[tool.ruff.lint.isort]
lines-after-imports = 2
no-sections = true

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["F704"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.pyright]
include = ["awslabs", "tests"]
exclude = ["**/__pycache__", "**/.venv", "**/node_modules", "**/dist", "**/build"]

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "live: marks tests that make live API calls (deselect with '-m \"not live\"')",
    "asyncio: marks tests that use asyncio"
]

[tool.coverage.report]
exclude_also = [
    'pragma: no cover',
    'if __name__ == .__main__.:\n    main()',
]

[tool.coverage.run]
source = ["awslabs"]
